requirements.txt
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.3.0
torch==2.4.0
torchvision==0.18.0
tqdm==4.66.4
